Refactor parameter norm calculations
Also includes improvements and fixes.
