+++ torchtitan/train.py
@@ -604,9 +604,16 @@ class Trainer(torch.distributed.checkpoint.stateful.Stateful):
             )
         self.checkpointer.maybe_wait_for_staging()
 
-        if self.job_config.metrics.log_norm_freq > 0 and (
-            self.step == 1 or self.step % self.job_config.metrics.log_norm_freq == 0
-        ):
+        # Here we let the optimizer know that we need to calculate the
+        # norm at the next step
+        need_to_calculate_norm = (
+            self.job_config.metrics.log_norm_freq > 0
+            and self.metrics_processor.should_log(self.step)
+            and (
+                self.step == 1 or self.step % self.job_config.metrics.log_norm_freq == 0
+            )
+        )
+        if need_to_calculate_norm:
             self.optimizers.calculate_norm_at_next_step()
 
         self.optimizers.step()
@@ -648,10 +655,14 @@ class Trainer(torch.distributed.checkpoint.stateful.Stateful):
             "lr": lr,
         }
         extra_metrics.update(self.optimizers.get_lrs())
-        if self.job_config.metrics.log_norm_freq > 0 and (
-            self.step == 1 or self.step % self.job_config.metrics.log_norm_freq == 0
-        ):
-            extra_metrics.update(self.optimizers.get_norms_at_current_step())
+
+        if need_to_calculate_norm:
+            # TODO(JSC)-Fixed: Notice that the original Gradient norm-log's LR is
+            #                  always one step behind the actual LR.
+            #                  Dist-scion can now calculate the norm at
+            #                  current step - synced LR.
+            param_norms = self.optimizers.get_parameter_norms()
+            extra_metrics.update(param_norms)
 
         if aux_loss is not None:
             extra_metrics["loss_metrics/aux_loss"] = aux_loss
